move some functions around in preparation for backend module
